@flar 2019-07-17
Newest update resolves:
- Expanded left/right/top/bottom += whxy logic to multi-line to fix analyzer issues
- Fixed variable names in test

I did not add any code to compare test results to vector_math